requirements.txt
pandas~=1.1.5
matplotlib~=3.3.3
numpy~=1.19.4
PyYAML~=5.3.1